// src/Api/models/AdminProfile.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type AdminProfile = {
    id?: string;
    name?: string;
    email?: string;
    phone?: string;
    role?: string;
    fromGoogle?: boolean;
    createdAt?: string;
    updatedAt?: string;
};